Extract prefabs for arena peices
To distinguish between types of walls, i added prefabs for horizontal and vertical walls, where horizontal ones have colliders for bouncing off of, and the left and right ones have no bouncing.

Of course, later on we will have to add removal of the ball upon hitting the goal line.