Add ai prediction for player ball trajectory - Part 2 balancing the ai (#38)
Finally got a more balanced ai that can handle arbitrary number of bounces, and is much cleaner than the previous part 1 implementation, as the details of the trajectory algorithm and coroutines were modularized (in class PredictedTrajectory and CoroutineUtils respectively).

Also, added a not so noticeable speed handicap to the AI, and a response time delay, to make it beatable (hence the balancing improvements). Few setting files changed to support this as a build, to.