Normalize ai behavior when defending ball and attempting to score (#91)
# Summary
Basically, this standardizes the AI behavior a bit more, mostly through changes in the BallTrajectoryPredictor, in which I tuned the optimal distances, tolerances, and other things related to predicting trajectories. Specifically, layers (ie Arena layer) are now taken into account by the ball predictor, and the max ray cast distance is set to be the vector length of a trajectory spanning all the way to the targetX position.

Additionally, I extracted out the paddlezones from the paddles and into children of a separate scene object. The reason for this was simple - the zones were actually moving dynamically with the paddles as their parent objects would be moving during gameplay.
Thus, by extracting it out, this effectively makes the PaddleZones encapsulate the paddles from top to bottom of the arena just like before, but now completely unmoving during playmode. In other words, the zones are now independent of any vertical movement, and instead the collider is updated when the paddle is moved in the editor, and at no other times.

Also made minor numeric adjustments to arena colliders, paddlezone widths, and paddle distance tolerances, all to smoothen the AI gameplay.


# Detailed commit log:
* Add raycast visual debug draw call if running in editor
* Simplify extrapolate point to use the entire position
* Relocate paddlezones so that their boxcolliders are independent of paddle movement
* Override paddlezone prefabs/paddles
* Replace tags and bounce data in TrajectoryPredictor class with layers and vertices
* Change gamesettings naming to match other primitive data classes
* Readjust paddle values for balancing reaction times and speed
* Remove redundant distance calculations

* Synchronize arena prefab to latest version of arena object
* Synchronize game scene to arena prefab changes
* Override all paddle and game prefab/variants to reflect game scene
* Realign arena prefabs to be consistent
* Rename `BallTrajectoryPredictor` `Compute` to `ComputeNewTrajectory` to provide more context to its side effects
* Readjust colliders to match arena mesh renderers
* First pass on correcting the distance of the raycast