Change percentage settings sent from main menu to be set and treated as ratios throughout rest of program, with mins and maxes for all (#94)
## The problem addressed
The issue this PR addresses the issue that when the difficulty, for example, is at lower (ie 20%) and higher difficulties (ie 80%) - the user experience is just ridiculous, as it doesn't make sense to have a paddle speed close to zero, for example. That's not just easy, its just useless.


## The solution
To change this, all those values are converted to float ratios with an inspector-assigned lower/upper bounds.

For example, we now have the fields `PADDLE_SPEED_AT_MIN_DIFFICULTY` `PADDLE_SPEED_AT_MAX_DIFFICULTY` assigned in the `AiPaddle`'s inspector panel, and everything is treated as a ratio corresponding to the percent chosen in the main menu, which corresponds to a value within the min and max range described above.

Similarly for the other slider values as well.